requirements.txt
torch
numpy-quaternion
scipy
tqdm
h5py
pandas
pathlib
pillow
pykeops
opencv-python
open3d
eigenpy
coloredlogs
nibabel
ruamel.yaml 
pytools
easydict
trimesh
sympy
torch_scatter
torch_cluster
pin